some tidying
